+++ NEWS.md
@@ -1,8 +1,12 @@
 
 3.10.3 (2015-mm-dd)
 ---
+* Cache public and with link embeds in redis [#3733](https://github.com/CartoDB/cartodb/pull/3733)
+
 Bugfixes:
 * Fixed deletion of layers upon disconnecting synced datasources [#3718](https://github.com/CartoDB/cartodb/pull/3718)
+* Fixed some cache invalidation problems upon changes in privacy (embeds & vizjson) [#3755](https://github.com/CartoDB/cartodb/pull/3755)
+* Fixed corner case with ghost table renames [#3762](https://github.com/CartoDB/cartodb/pull/3762)
 
 3.10.2 (2015-05-20)
 ---------
@@ -23,6 +27,7 @@ git submodule update --init
 
 Bugfixes:
 * Create import view throws uncaught error on canceling selected a file (2nd time) [#3379](https://github.com/CartoDB/cartodb/issues/3379).
+* Removed old tags endpoint (https://github.com/CartoDB/cartodb/issues/3691).
 
 3.10.1 (2015-05-05)
 -------------------